[NEAR-119] Refactor: Concert 서비스 로직 분리
- concert를 stream에서 독립
