+++ app/domains/concerts/service.py
@@ -1,0 +1,140 @@
+from typing import Any
+
+from fastapi import HTTPException, UploadFile
+
+from app.core.pagination import paginate_cursor
+from app.core.utils.image_resizer import ImageResizer
+from app.core.utils.permissions import AdminPermission
+from app.domains.concerts.models import Concert
+from app.domains.concerts.schemas import (
+    ConcertCreateRequest,
+    ConcertDetailResponse,
+)
+from app.domains.users.models import User
+from app.integrations.aws_ivs import IVSClient
+from app.integrations.aws_ivs.client import logger
+
+
+class ConcertAdminService:
+    def __init__(self, ivs_client: IVSClient):
+        self.ivs_client = ivs_client
+        self.image_resizer = ImageResizer()
+
+    async def create_concert(self, data: ConcertCreateRequest, user: User) -> Concert:
+        """
+        [Admin] 콘서트 생성
+        """
+        AdminPermission.must_be_admin(user)
+        return await Concert.create(**data.model_dump())
+
+    async def update_concert_thumbnail(
+        self, concert_id: int, file: UploadFile, user: User
+    ) -> Concert:
+        """
+        콘서트 썸네일 생성 및 수정
+        """
+        AdminPermission.must_be_admin(user)
+
+        # 콘서트 있는지 확인
+        concert = await Concert.get_or_none(id=concert_id)
+        if not concert:
+            raise HTTPException(status_code=404, detail="콘서트를 찾을 수 없습니다.")
+
+        # 기존 이미지가 있다면 S3 폴더 삭제
+        if concert.thumbnail_url:
+            await self.image_resizer.delete_all_by_id_path(concert.thumbnail_url)
+
+        # 새로운 이미지 리사이징 업로드
+        path_prefix = f"concerts/{concert.id}/thumbnail"
+        sizes = (300, 640, 1280)
+
+        urls = await self.image_resizer.upload_square_resizes(
+            image_file=file.file, sizes=sizes, path_prefix=path_prefix
+        )
+
+        # DB 업데이트
+        img_url = urls.get("640")
+        concert.thumbnail_url = img_url if img_url else ""
+        await concert.save()
+
+        return concert
+
+    async def list_concerts(
+        self, user: User, limit: int = 20, cursor: int | None = None
+    ) -> tuple[list[Concert], int | None]:
+        """
+        콘서트 목록 조회
+        """
+        AdminPermission.must_be_admin(user)
+
+        queryset = Concert.all()
+        result: Any = await paginate_cursor(queryset, limit=limit, cursor=cursor, order_by="-id")
+
+        concerts: list[Concert] = result[0]
+        next_cursor: int | None = result[1]
+
+        return concerts, next_cursor
+
+    async def get_concert_detail(self, concert_id: int, user: User) -> ConcertDetailResponse:
+        """
+        콘서트 상세 조회(하위 세션 목록 포함)
+        """
+        AdminPermission.must_be_admin(user)
+
+        concert = await Concert.get_or_none(id=concert_id).prefetch_related(
+            "sessions__stream_channel"
+        )
+
+        if not concert:
+            raise HTTPException(404, "존재하지 않는 콘서트입니다.")
+
+        return ConcertDetailResponse.model_validate(concert)
+
+    async def update_concert(
+        self, concert_id: int, data: ConcertCreateRequest, user: User
+    ) -> Concert:
+        """
+        콘서트 수정
+        """
+        AdminPermission.must_be_admin(user)
+
+        concert = await Concert.get_or_none(id=concert_id)
+        if not concert:
+            raise HTTPException(404, "수정할 콘서트를 찾을 수 없습니다.")
+
+        # 요청 데이터 반영 (exclude_unset=True로 보낸 값만 수정)
+        update_data = data.model_dump(exclude_unset=True)
+        for key, value in update_data.items():
+            setattr(concert, key, value)
+
+        await concert.save()
+        return concert
+
+    async def delete_concert_with_infrastructure(self, concert_id: int, user: User) -> None:
+        """
+        콘서트 삭제 및 AWS IVS 채널 삭제
+        """
+        AdminPermission.must_be_admin(user)
+
+        # 세션, 채널 정보 모두 가져옴
+        concert = await Concert.get_or_none(id=concert_id).prefetch_related(
+            "sessions__stream_channel"
+        )
+
+        if not concert:
+            raise HTTPException(404, "삭제할 콘서트를 찾을 수 없습니다.")
+
+        # 연결된 모든 세션의 AWS IVS 채널 먼저 삭제
+        for session in list(concert.sessions):  # type: ignore[call-overload]
+            if hasattr(session, "stream_channel") and session.stream_channel:
+                try:
+                    self.ivs_client.delete_channel(session.stream_channel.channel_arn)
+                except Exception as e:
+                    logger.warning(f"콘서트 삭제 중 세션({session.id})의 IVS 채널 삭제 실패: {e}")
+
+        # 썸네일 삭제 (S3)
+        if concert.thumbnail_url:
+            await self.image_resizer.delete_all_by_id_path(concert.thumbnail_url)
+
+        # DB 삭제(cascade)
+        await concert.delete()
